so yeah about unit test
they kinda work if test individual inside the code but as it test
like create a proper test file and to use coverage package to find
test coverage this is very very diffucult? what? i mean this looming
probleming for a while as expect that this will bite my arse
anyways i think i can find the way to solve it on time